Makes unrolling tree a DAG.
The idea of this commit is to recycle b and e variables in function
template instantiations. This essentially means that the graph of guards
(b variables), which used to be a tree, is now a DAG. The hope is that
this limits the number of unrollings, as different instantiations with
the same arguments are now unrolled only once.