imperative testcases for scala-workshop 2013
Based on testcases from vstte13 and functional benchmarks
of SAS. We convert some of the functional benchmarks to an
equivalent imperative version.

We are not able to reproduce all successfull properties of
the functional benchmarks, especially when the function we
are implementation was not originally tail recursive. In
that case, a non trivial encoding is required.

Insertion sort and other sort algorithms are particularly
complicated to implement with an imperative style. Function
like insert need to use a reverse while reconstructing the
list, and need in particular to prove reverse of an increasing
list is a decreasing list. We are not able to prove that so far.