Update GenC regression tests
Tests that couldn't be verified are now in a "unverified" folder: they are not verified with the GenC regression test but they are still converted to C and their execution should yield 0.

At the moment, in ExpressionOrder, test7 fails to verify. However, it should pass because my Scala interpreter successfully return true.