Add GenC v2
This new implementation of GenC uses a pipelined approach: the translation of Scala code into C99 equivalent code is split into several phases that live in the `leon.genc.phases` package. The overall process is bases on several versions of an Intermediate Representation (IR) that is defined under `leon.genc.ir`. Additionally, the C AST was partially redesigned.

This new implementation is supposed to be better than the previous version in terms of bugs but there remain a few areas for improvement:
  * The `ClassLifter` is slightly incomplete: class fields should be processed as well.
  * When converting nested generic functions, the Scala2IRPhase might not do a proper job.

Both of these points are explained in the source code; see the corresponding TODOs.